+++ client/src/main/java/eu/cqse/teamscale/jacoco/client/report/XmlReportGenerator.java
@@ -1,22 +1,16 @@
-package eu.cqse.teamscale.jacoco.client;
+package eu.cqse.teamscale.jacoco.client.report;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Predicate;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
 import org.jacoco.core.analysis.Analyzer;
 import org.jacoco.core.analysis.CoverageBuilder;
 import org.jacoco.core.analysis.IBundleCoverage;
-import org.jacoco.core.analysis.IClassCoverage;
 import org.jacoco.core.data.ExecutionDataStore;
 import org.jacoco.report.IReportVisitor;
 import org.jacoco.report.xml.XMLFormatter;
@@ -33,10 +27,7 @@ public class XmlReportGenerator {
 	/**
 	 * Include filter to apply to all locations during class file traversal.
 	 */
-	private final Predicate<Path> locationIncludeFilter;
-
-	/** The logger. */
-	private final Logger logger = LogManager.getLogger(this);
+	private final Predicate<String> locationIncludeFilter;
 
 	/** Whether to ignore non-identical duplicates of class files. */
 	private final boolean ignoreNonidenticalDuplicateClassFiles;
@@ -44,7 +35,7 @@ public class XmlReportGenerator {
 	/**
 	 * Constructor.
 	 */
-	public XmlReportGenerator(List<File> codeDirectoriesOrArchives, Predicate<Path> locationIncludeFilter,
+	public XmlReportGenerator(List<File> codeDirectoriesOrArchives, Predicate<String> locationIncludeFilter,
 			boolean ignoreDuplicates) {
 		this.codeDirectoriesOrArchives = codeDirectoriesOrArchives;
 		this.ignoreNonidenticalDuplicateClassFiles = ignoreDuplicates;
@@ -85,42 +76,13 @@ public class XmlReportGenerator {
 			coverageBuilder = new DuplicateIgnoringCoverageBuilder();
 		}
 
-		Analyzer analyzer = new Analyzer(store, coverageBuilder) {
-			@Override
-			public int analyzeAll(InputStream input, String location) throws IOException {
-				if (location.endsWith(".class") && !locationIncludeFilter.test(Paths.get(location))) {
-					logger.debug("Filtering class file {}", location);
-					return 0;
-				}
-				return super.analyzeAll(input, location);
-			}
-		};
+		Analyzer analyzer = new FilteringAnalyzer(store, coverageBuilder, locationIncludeFilter);
 
 		for (File file : codeDirectoriesOrArchives) {
 			analyzer.analyzeAll(file);
 		}
 
 		return coverageBuilder.getBundle("dummybundle");
-	}
-
-	/** Modified {@link CoverageBuilder} that ignores non-identical duplicates. */
-	private static class DuplicateIgnoringCoverageBuilder extends CoverageBuilder {
-
-		/** The logger. */
-		private final Logger logger = LogManager.getLogger(this);
-
-		/**
-		 * {@inheritDoc}
-		 */
-		@Override
-		public void visitCoverage(IClassCoverage coverage) {
-			try {
-				super.visitCoverage(coverage);
-			} catch (IllegalStateException e) {
-				logger.warn("Ignoring duplicate, non-identical class file for class {}", coverage.getName(), e);
-			}
-		}
-
 	}
 
 }