+++ client/src/main/java/eu/cqse/teamscale/jacoco/client/watch/Watcher.java
@@ -20,7 +20,7 @@ import org.apache.logging.log4j.MarkerManager;
 import org.conqat.lib.commons.collections.CollectionUtils;
 import org.conqat.lib.commons.string.StringUtils;
 
-import eu.cqse.teamscale.jacoco.client.XmlReportGenerator;
+import eu.cqse.teamscale.jacoco.client.report.XmlReportGenerator;
 import eu.cqse.teamscale.jacoco.client.store.IXmlStore;
 import eu.cqse.teamscale.jacoco.client.store.TimestampedFileStore;
 import eu.cqse.teamscale.jacoco.client.util.AntPatternIncludeFilter;